+++ web-wallet/CHANGELOG.md
@@ -12,13 +12,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Update Transactions list design [#1922]
-- Change password setting as enabled by default [#3287]
+
+### Removed
+
+### Fixed
+
+## [1.1.0] - 2025-01-09
+
+### Changed
+
+- Make Password setting enabled by default (Create/Restore Wallet flow) [#3287]
 
 ### Removed
 
 - Remove "Swap to nDUSK" step (Create Wallet flow) [#3333]
-
-### Fixed
 
 ## [1.0.0] - 2024-12-23
 
@@ -510,6 +517,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- VERSIONS -->
 
 [Unreleased]: https://github.com/dusk-network/rusk/tree/master/web-wallet
+[1.1.0]: https://github.com/dusk-network/rusk/tree/web-wallet-v1.0.1
 [1.0.0]: https://github.com/dusk-network/rusk/tree/web-wallet-v1.0.0
 [0.10.1]: https://github.com/dusk-network/rusk/tree/web-wallet-v0.10.1
 [0.10.0]: https://github.com/dusk-network/rusk/tree/web-wallet-v0.10.0